Time: 159 ms (29.82%) | Memory: 79.2 MB (20.00%) - LeetSync

+++ HashTable/1995-finding-pairs-with-a-certain-sum/finding-pairs-with-a-certain-sum.java
@@ -1,7 +1,7 @@
 class FindSumPairs {
 
-    Map<Long, Long> map1 = null;
-    Map<Long, Long> map2 = null;
+    Map<Integer, Integer> map1 = null;
+    Map<Integer, Integer> map2 = null;
     int[] nums1;
     int[] nums2;
 
@@ -14,15 +14,15 @@ class FindSumPairs {
     public FindSumPairs(int[] nums1, int[] nums2) {
         this.nums1 = nums1;
         this.nums2 = nums2;
-        map1 = new HashMap<Long, Long>(); // SC: O(N1)
-        map2 = new HashMap<Long, Long>(); // SC: O(N2)
+        map1 = new HashMap<Integer, Integer>(); // SC: O(N1)
+        map2 = new HashMap<Integer, Integer>(); // SC: O(N2)
         int n1 = nums1.length;
         int n2 = nums2.length;
         for (int i = 0; i < n1; i++) { // TC: O(N1)
-            map1.put((long) nums1[i], map1.getOrDefault((long) nums1[i], 0L) + 1);
+            map1.put(nums1[i], map1.getOrDefault(nums1[i], 0) + 1);
         }
         for (int i = 0; i < n2; i++) { // TC: O(N2)
-            map2.put((long) nums2[i], map2.getOrDefault((long) nums2[i], 0L) + 1);
+            map2.put(nums2[i], map2.getOrDefault(nums2[i], 0) + 1);
         }
     }
     
@@ -33,14 +33,14 @@ class FindSumPairs {
      * SC: O(1)
      */
     public void add(int index, int val) {
-        long oldValue = (long) nums2[index];
-        map2.put(oldValue, map2.getOrDefault(oldValue, 0L) - 1);
+        int oldValue = nums2[index];
+        map2.put(oldValue, map2.getOrDefault(oldValue, 0) - 1);
         if (map2.get(oldValue) == 0) {
             map2.remove(oldValue);
         }
-        long newValue = oldValue + val;
-        nums2[index] = (int) newValue;
-        map2.put(newValue, map2.getOrDefault(newValue, 0L) + 1);
+        int newValue = oldValue + val;
+        nums2[index] = newValue;
+        map2.put(newValue, map2.getOrDefault(newValue, 0) + 1);
     }
     
     /**
@@ -50,16 +50,16 @@ class FindSumPairs {
      * SC: O(1)
      */
     public int count(int tot) {
-        long pairs = 0;
-        for (Long key : map1.keySet()) { // TC: O(N1)
-            long diff = (long) tot - key;
+        int pairs = 0;
+        for (Integer key : map1.keySet()) { // TC: O(N1)
+            int diff = tot - key;
             if (map2.containsKey(diff)) {
-                long freq1 = map1.get(key);
-                long freq2 = map2.get(diff);
+                int freq1 = map1.get(key);
+                int freq2 = map2.get(diff);
                 pairs += (freq1 * freq2);
             }
         }
-        return (int) pairs;
+        return pairs;
     }
 }
